requirements.txt
Django==4.1.3
django-compressor
psycopg2-binary
djangorestframework
tzdata